+++ substrafl/algorithms/pytorch/torch_scaffold_algo.py
@@ -105,10 +105,6 @@ class TorchScaffoldAlgo(TorchAlgo):
                         self._scaffold_parameters_update()
                         if self._scheduler is not None:
                             self._scheduler.step()
-
-    As development tools, the ``train`` and ``predict`` method comes with a default argument : ``_skip``.
-    If ``_skip`` is set to ``True``, only the function will be executed and not all the code related to Substra.
-    This allows to quickly debug code and use the defined algorithm as is.
     """
 
     def __init__(